Implement rate limiting and settings bottom sheet
Implement rate limiting and feed "modes" - currently, two: a rate
limited mode, or a mode where messages are provided if the sentiment is
below a certain threshold.

Move theme settings to a new modal bottom sheet, now that we have
additional configuration options.

Implement UI elements to allow selecting feed modes, and to set the rate
limit and threshold values, when applicable.

Implement command parsing on the server side to enable the client to
instruct the server on mode/rate/threshold settings.

Implement per-client buffering of messages on the server side, to allow
clients to set a threshold or rate that does not impact other clients.

Implement a new "hatestack" type for the server, which is designed to
keep only the worst sentiment messages as they are pushed onto a stack
of fixed size.

Bump version of client and server to v0.3.0 to reflect new features.