v0.4.0 Implement profile resolution
daemon: Implement profile resolution, which resolves the DIDs the
firehose provides into a display name and handle. Create a cache (using
hashicorp/golang-lru) for storing these to reduce re-resolution of the
DIDs.

daemon: Add display name to the ProcessedPost data model. Use pointers
so that they can be instantiated as nil then filled in JIT by feed
fanout receivers before sending to clients. Profiles are resolved JIT so
we don't hammer the BlueSky API to resolve everything coming across the
firehose when it may never be shown to the client.

ui: Add display name above the handle in the post card, and place handle
below it in a smaller font. Add displayName to the ProcessedPost
constructor / parser.