Add sentiment coloring for individual words
Add a toggle for showing sentiment coloring on individual words in the
post body. Very negative sentiment will color the word red and bold it,
and will ease towards the regular text color / font weight for less
intense scores.

Sentiment analysis is performed server-side and passed to the client as
a list of pre-scored tokens. Breaks protocol compatibility with
frontends < v0.5.